+++ api/src/main/java/com/basistech/util/TextDomain.java
@@ -17,11 +17,12 @@
 package com.basistech.util;
 
 import java.io.Serializable;
+import java.util.Objects;
 
 /**
  * Collection of linguistic properties of a text which are independent of its message; this currently includes
  * script, language, and transliteration scheme. Any text is in a domain (with zero or more components under
- * specified). Instances of this class are immutable.
+ * specified).
  */
 public class TextDomain implements Serializable, Comparable<TextDomain> {
 
@@ -31,30 +32,36 @@ public class TextDomain implements Serializable, Comparable<TextDomain> {
 
     /**
      * Create a TextDomain object.
+     * {@code null} values supplied here are mapped to their corresponding defaults.
      * 
-     * @param script the ISO 15924 numeric script id
-     * @param language the ISO 639 numeric langauge id
-     * @param scheme the TransliterationScheme
+     * @param script the ISO 15924 code. {@code null} is mapped to {@link ISO15924#Zyyy}.
+     * @param language the language code. {@code null} is mapped to {@link LanguageCode#UNKNOWN}.
+     * @param scheme the TransliterationScheme. {@code null} is mapped to {@link TransliterationScheme#UNKNOWN}.
      */
     public TextDomain(ISO15924 script, LanguageCode language, TransliterationScheme scheme) {
-        theScript = script;
-        theLanguage = language;
-        theScheme = scheme;
+        theScript = script == null ? ISO15924.Zyyy : script;
+        theLanguage = language == null ? LanguageCode.UNKNOWN : language;
+        theScheme = scheme == null ? TransliterationScheme.UNKNOWN : scheme;
     }
 
     /**
-     * Create a TextDomain object for the "native" domain of the language (see
-     * LanguageCode.getDefaulatScript())
+     * Create a TextDomain object for the default domain of the language (see
+     * LanguageCode.getDefaultScript())
      * 
-     * @param language the ISO 639 numeric langauge id
+     * @param language the language.
+     *                 {@code null} is mapped to {@link LanguageCode#UNKNOWN}.
      */
     public TextDomain(LanguageCode language) {
+        if (language == null) {
+            language = LanguageCode.UNKNOWN;
+        }
         theScript = language.getDefaultScript();
         theLanguage = language;
-        theScheme = TransliterationScheme.NATIVE;
+        theScheme = TransliterationScheme.UNKNOWN;
     }
     /**
-     * Create a TextDomain object 
+     * Create a TextDomain object with
+     * unspecified values.
      * 
      */
     TextDomain() {
@@ -63,33 +70,23 @@ public class TextDomain implements Serializable, Comparable<TextDomain> {
         theScheme = TransliterationScheme.UNKNOWN;
     }
 
-    /**
-     * Determine if the contents of this TextDomain object are equal those in the given object.
-     * 
-     * @param o object to compare
-     * @return true if contents are equal.
-     */
+    @Override
     public boolean equals(Object o) {
-        if (!(o instanceof TextDomain)) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        return this.getScript() == ((TextDomain)o).getScript()
-               && this.getLanguage() == ((TextDomain)o).getLanguage()
-               && this.getTransliterationScheme() == ((TextDomain)o).getTransliterationScheme();
-
+        TextDomain that = (TextDomain) o;
+        return theScript == that.theScript
+                && theLanguage == that.theLanguage
+                && theScheme == that.theScheme;
     }
 
-    /**
-     * Returns the hash code for this TextDomain.
-     * 
-     * @return hash code
-     */
+    @Override
     public int hashCode() {
-        int result = 17;
-        result = 37 * result + this.getScript().hashCode();
-        result = 37 * result + this.getLanguage().hashCode();
-        result = 37 * result + this.getTransliterationScheme().hashCode();
-        return result;
+        return Objects.hash(theScript, theLanguage, theScheme);
     }
 
     /**
@@ -103,10 +100,13 @@ public class TextDomain implements Serializable, Comparable<TextDomain> {
 
     /**
      * Set the script of this domain.
-     * 
+     * @param script the new script value.
+     *     {@code null} is mapped to {@link ISO15924#Zyyy}.
+     * @deprecated This method will be removed in a future version to allow these objects to be immutable.
      */
+    @Deprecated
     void setScript(ISO15924 script) {
-        theScript = script;
+        theScript = script == null ? ISO15924.Zyyy : script;
     }
 
     /**
@@ -120,10 +120,14 @@ public class TextDomain implements Serializable, Comparable<TextDomain> {
 
     /**
      * Set the language of this domain.
+     * @param lang the new language value.
+     *             {@code null} is mapped to {@link LanguageCode#UNKNOWN}.
+     * @deprecated This method will be removed in a future version to allow these objects to be immutable.
      * 
      */
+    @Deprecated
     void setLanguage(LanguageCode lang) {
-        theLanguage = lang;
+        theLanguage = lang == null ? LanguageCode.UNKNOWN : lang;
     }
  
     /**
@@ -137,10 +141,13 @@ public class TextDomain implements Serializable, Comparable<TextDomain> {
 
     /**
      * set the transliteration scheme of this domain.
-     * 
+     * @param scheme the new scheme value.
+     *               {@code null} is mapped to {@link TransliterationScheme#UNKNOWN}.
+     * @deprecated This method will be removed in a future version to allow these objects to be immutable.
      */
+    @Deprecated
     void setTransliterationScheme(TransliterationScheme scheme) {
-        theScheme = scheme;
+        theScheme = scheme == null ? TransliterationScheme.UNKNOWN : scheme;
     }
 
     /**
@@ -148,30 +155,46 @@ public class TextDomain implements Serializable, Comparable<TextDomain> {
      * 
      * @return String representation of the TextDomain.
      */
+    @Override
     public String toString() {
-        return "[" + theScript.code4() + "/" + theLanguage.ISO639_3() + "/" + theScheme.getName() + "]";
+        return "["
+                + theScript.code4()
+                + "/"
+                + theLanguage.ISO639_3()
+                + "/"
+                + theScheme.getName()
+                + "]";
     }
 
-    /**
-     * {@inheritDoc}
-     */
+    @Override
     public int compareTo(TextDomain o) {
-        if (theScript.numeric() > o.getScript().numeric()) {
+        int n = theScript.numeric();
+        int otherN = o.getScript().numeric();
+
+        if (n > otherN) {
             return 1;
-        } else if (theScript.numeric() < o.getScript().numeric()) {
+        } else if (n < otherN) {
             return -1;
         }
-        if (theLanguage.languageID() > o.getLanguage().languageID()) {
+
+        n = theLanguage.languageID();
+        otherN = o.getLanguage().languageID();
+
+        if (n > otherN) {
             return 1;
-        } else if (theLanguage.languageID() < o.getLanguage().languageID()) {
+        } else if (n < otherN) {
             return -1;
         }
-        if (theScheme.getNativeCode() > o.getTransliterationScheme().getNativeCode()) {
+
+        n = theScheme.getNativeCode();
+        otherN = o.getTransliterationScheme().getNativeCode();
+
+        if (n > otherN) {
             return 1;
-        } else if (theScheme.getNativeCode() < o.getTransliterationScheme().getNativeCode()) {
+        } else if (n < otherN) {
             return -1;
         }
- 
+
         return 0;
     }
 }